requirements.txt
appdirs==1.4.3
backports-abc==0.5
certifi==2017.4.17
evdev==1.2.0
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-speech==1.2.0
google-speech==1.1.0
googleapis-common-protos==1.6.0
gpiozero==1.5.1
grpcio==1.22.0
matrix-io-proto==0.0.10
packaging==16.8
protobuf==3.9.0
pyparsing==2.2.0
pyzmq==17.1.2
singledispatch==3.4.0.3
six==1.10.0
tornado==5.1.1
gobject==0.1.0
matrix-lite==0.0.3
